requirements.txt
black
duckdb==1.2.2
fastapi==0.115.14
flake8
httpx==0.28.1
mkdocs
mkdocs-material
mkdocs-mermaid2-plugin
pre-commit
psycopg2-binary==2.9.11
pytest
pytest-asyncio==0.25.3
strawberry-graphql[fastapi]==0.272.1
uvicorn==0.34.3